+++ projects/ui/src/lib/components/index.ts
@@ -38,3 +38,4 @@ export * from './stepper';
 export * from './breadcrumbs';
 export * from './calendar';
 export * from './date-picker';
+export * from './tree';