+++ projects/ui/src/lib/components/tree/index.ts
@@ -1,0 +1,2 @@
+export { KuiTreeComponent } from './kui-tree.component';
+export type { KuiTreeMode, KuiTreeNode } from './kui-tree-node.interface';